Rename mailchimpApiConfig to more clearly state purpose
